Introduce Voice namespace and voice_client to access voice API
