+++ lib/messagebird.rb
@@ -17,4 +17,6 @@ require 'messagebird/signed_request'
 require 'messagebird/verify'
 require 'messagebird/message'
 require 'messagebird/voicemessage'
-require 'messagebird/call'
+require 'messagebird/voice/call'
+require 'messagebird/voice/call_leg'
+require 'messagebird/voice/call_leg_recording'
