+++ lib/messagebird/client.rb
@@ -17,8 +17,11 @@ require 'messagebird/lookup'
 require 'messagebird/message'
 require 'messagebird/verify'
 require 'messagebird/voicemessage'
-require 'messagebird/call'
-require 'messagebird/call/list'
+require 'messagebird/voice_client'
+require 'messagebird/voice/call'
+require 'messagebird/voice/call_leg'
+require 'messagebird/voice/call_leg_recording'
+require 'messagebird/voice/list'
 
 module MessageBird
   class ErrorException < StandardError
@@ -33,16 +36,17 @@ module MessageBird
   end
 
   class Client
-    attr_reader :access_key, :http_client, :conversation_client
+    attr_reader :access_key, :http_client, :conversation_client, :voice_client
 
     CONVERSATIONS_WHATSAPP_SANDBOX_FEATURE = "CONVERSATIONS_WHATSAPP_SANDBOX_FEATURE" # Enables the whatsapp sandbox
     VALID_FEATURES = [CONVERSATIONS_WHATSAPP_SANDBOX_FEATURE]                         # List of valid features for validation
 
-    def initialize(access_key = nil, http_client = nil, conversation_client = nil)
+    def initialize(access_key = nil, http_client = nil, conversation_client = nil, voice_client = nil)
       @access_key = access_key || ENV['MESSAGEBIRD_ACCESS_KEY']
       @http_client = http_client || HttpClient.new(@access_key)
 
       @conversation_client = conversation_client || ConversationClient.new(@access_key)
+      @voice_client = voice_client || VoiceClient.new(@access_key)
     end
 
     def enable_feature(feature)
@@ -59,6 +63,13 @@ module MessageBird
       else
         raise InvalidFeatureException
       end
+    end
+
+    def voice_request(method, path, params={})
+      response_body = @voice_client.request(method, path, params)
+    
+      return if response_body.nil? || response_body.empty?
+      parse_body(response_body)
     end
 
     def conversation_request(method, path, params={})
@@ -240,22 +251,23 @@ module MessageBird
     def call_create(source, destination, call_flow = {}, webhook = {}, params={})
       params = params.merge({callFlow: call_flow.to_json}) unless call_flow.empty?
       params = params.merge({webhook: webhook.to_json}) unless webhook.empty?
-      Call.new(request(
+      Voice::Call.new(voice_request(
         :post,
         'calls',
         params.merge({ :source => source, :destination => destination })))
     end
 
-    def call_list(per_page = CallList::PER_PAGE, page = CallList::CURRENT_PAGE)
-      CallList.new(Call, request(:get, "calls?perPage=#{per_page}&page=#{page}"))
+    def call_list(per_page = Voice::List::PER_PAGE, page = Voice::List::CURRENT_PAGE)
+      Voice::List.new(Voice::Call, voice_request(:get, "calls?perPage=#{per_page}&currentPage=#{page}"))
     end
 
     def call_view(id)
-      Call.new(request(:get, "calls/#{id.to_s}"))
+      Voice::Call.new(voice_request(:get, "calls/#{id.to_s}"))
     end
 
     def call_delete(id)
-      request(:delete, "calls/#{id}")
+      voice_request(:delete, "calls/#{id}")
+    end
     end
 
     def lookup(phoneNumber, params={})
